reversal of the DTOs renaming
