+++ .gitignore
@@ -31,3 +31,4 @@ docker-compose.override.yml
 
 /docker/traefik/.env.example
 /docker/pgadmin
+/docker/traefik/certs
