+++ app/Http/Controllers/API/WpOrg/Plugins/PluginUpdateCheck_1_1_Controller.php
@@ -4,7 +4,7 @@ namespace App\Http\Controllers\API\WpOrg\Plugins;
 
 use App\Http\Controllers\Controller;
 use App\Services\PluginServices\PluginUpdateService;
-use App\Values\WpOrg\PluginDTOs\PluginUpdateCheckRequest;
+use App\Values\WpOrg\Plugins\PluginUpdateCheckRequest;
 use Illuminate\Http\JsonResponse;
 use Illuminate\Http\Request;
