+++ app/Services/PluginServices/PluginUpdateService.php
@@ -3,9 +3,9 @@
 namespace App\Services\PluginServices;
 
 use App\Models\WpOrg\Plugin;
-use App\Values\WpOrg\PluginDTOs\PluginUpdateCheckRequest;
-use App\Values\WpOrg\PluginDTOs\PluginUpdateCheckResponse;
-use App\Values\WpOrg\PluginDTOs\PluginUpdateData;
+use App\Values\WpOrg\Plugins\PluginUpdateCheckRequest;
+use App\Values\WpOrg\Plugins\PluginUpdateCheckResponse;
+use App\Values\WpOrg\Plugins\PluginUpdateData;
 
 class PluginUpdateService
 {
