+++ app/Values/WpOrg/Plugins/ClosedPluginResponse.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace App\Values\WpOrg\PluginDTOs;
+namespace App\Values\WpOrg\Plugins;
 
 use App\Models\WpOrg\ClosedPlugin;
 use App\Values\DTO;
