+++ app/Values/WpOrg/Plugins/PluginUpdateData.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Values\WpOrg\PluginDTOs;
+namespace App\Values\WpOrg\Plugins;
 
 use App\Models\WpOrg\Plugin;
 use App\Values\DTO;
