+++ app/Values/WpOrg/Plugins/QueryPluginsRequest.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace App\Values\WpOrg\PluginDTOs;
+namespace App\Values\WpOrg\Plugins;
 
 use App\Utils\Regex;
 use App\Values\DTO;
@@ -15,7 +15,7 @@ use Illuminate\Support\Arr;
 // I'd look into refactoring it, but it's not like .org is going to add a new resource type anytime soon.
 // We can clean things up in the 2.0 API.
 #[StripExtraParameters]
-readonly class QueryPluginsDTO extends DTO
+readonly class QueryPluginsRequest extends DTO
 {
     /** @param list<string>|null $tags */
     public function __construct(
